Move quasiquoter's internals to own file
This will make them available for other modules
