+++ pirouette.cabal
@@ -18,6 +18,7 @@ library
       Language.Pirouette.Example
       Language.Pirouette.Example.IsUnity
       Language.Pirouette.QuasiQuoter
+      Language.Pirouette.QuasiQuoter.Internal
       Language.Pirouette.QuasiQuoter.Syntax
       Language.Pirouette.QuasiQuoter.ToTerm
       ListT.Weighted
